+++ packages/webview_flutter/webview_flutter_platform_interface/lib/src/types/navigation_request.dart
@@ -5,11 +5,18 @@
 /// Defines the parameters of the pending navigation callback.
 class NavigationRequest {
   /// Creates a [NavigationRequest].
-  const NavigationRequest({required this.url, required this.isMainFrame});
+  const NavigationRequest({required this.url, required this.isMainFrame, this.headers});
 
   /// The URL of the pending navigation request.
   final String url;
 
   /// Indicates whether the request was made in the web site's main frame or a subframe.
   final bool isMainFrame;
+
+  /// Headers for the request.
+  final Map<String, String>? headers;
+
+  /// Indicates whether the request was made to download an attachment.
+  bool get isDownloadRequest =>
+      headers?['content-disposition']?.toLowerCase().startsWith('attachment') ?? false;
 }